뒤로가기 발생 시 state reset 로직을 커스텀훅으로 작성
